+++ xkeymacsdll/xkeymacsdll.h
@@ -139,7 +139,7 @@ private:
 	static int CheckOriginal(UINT nType, BYTE bVk);
 	static void InitKeyboardProc(BOOL bImeComposition);
 	static int m_nAppID;
-	static int m_nOriginal[MAX_COMMAND_TYPE][MAX_KEY];
+	static BYTE m_nOriginal[MAX_COMMAND_TYPE][MAX_KEY];
 	static int m_nKillRing;
 	static CList<CClipboardSnap *, CClipboardSnap *> m_oKillRing;
 	static BOOL m_bHook;
